+++ src/main/java/com/hamoyeah/contents/repository/ShopRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ShopRepository extends JpaRepository<ShopEntity, Long> {
+public interface ShopRepository extends JpaRepository<ShopEntity, Integer> {
     List<ShopEntity> findByShopCategory(ShopCategory shopCategory); // 카테고리 분류를 리스트화
 }
